Break off `CombinatorTests` from `AdvancedTests`
